+++ include/rcppsw/ds/base_overlay_grid2D.hpp
@@ -72,15 +72,15 @@ class base_overlay_grid2D : public base_grid2D<T> {
    * @brief Get the size of the X dimension of the discretized subgrid, at
    * whatever the resolution specified during object construction was.
    */
-  uint xdsize(void) const {
-    return static_cast<uint>(std::ceil(m_x_max / m_resolution));
+  size_t xdsize(void) const {
+    return static_cast<size_t>(std::ceil(m_x_max / m_resolution));
   }
   /**
    * @brief Get the size of the Y dimension of the discretized subgrid, at
    * whatever the resolution specified during object construction was.
    */
-  uint ydsize(void) const {
-    return static_cast<uint>(std::ceil(m_y_max / m_resolution));
+  size_t ydsize(void) const {
+    return static_cast<size_t>(std::ceil(m_y_max / m_resolution));
   }
 
   /**
@@ -104,7 +104,7 @@ class base_overlay_grid2D : public base_grid2D<T> {
    * @return A pair containing the upper/lower X coordinates of the circle when
    * applied to the specified X coordinate.
    */
-  circle_range_ret_type circle_xrange_at_point(uint x, uint radius) const {
+  circle_range_ret_type circle_xrange_at_point(size_t x, uint radius) const {
     typename index_range::index lower_x =
         static_cast<typename index_range::index>(
             std::max<int>(0, static_cast<int>(x) - static_cast<int>(radius)));
@@ -128,7 +128,7 @@ class base_overlay_grid2D : public base_grid2D<T> {
    * @return A pair containing the upper/lower Y coordinates of the circle when
    * applied to the specified Y coordinate.
    */
-  circle_range_ret_type circle_yrange_at_point(uint y, uint radius) const {
+  circle_range_ret_type circle_yrange_at_point(size_t y, uint radius) const {
     typename index_range::index lower_y =
         static_cast<typename index_range::index>(
             std::max<int>(static_cast<int>(0),
